refactor: ServiceIntegrationTest로 네이밍 수정
